- Добавлены сеттеры на проверку длин сторон
- Исправлена документация
