+++ ЯВУ 3/ЯВУ 3/Pair.cpp
@@ -1,4 +1,3 @@
-#include<iostream>
 #include"Pair.h"
 using namespace std;
 Pair::Pair() 
@@ -11,31 +10,14 @@ Pair::Pair(double a, double b)
 		this->a = a;
 		this->b = b;
 	}
+void Pair::set_changing_a(double change_a)
+{
+	a = change_a;
+}
+void Pair::set_changing_b(double change_b)
+{
+	b = change_b;
+}
 double Pair::get_a() { return a; }
 double Pair::get_b() { return b; }
 double Pair::get_composition() { return a * b; }
-void Pair::set_changing_numbers(double change_a, double change_b) 
-{
-	a = change_a;
-	b = change_b;
-}
-double Pair::get_changed_number_a() { return a; }
-double Pair::get_changed_number_b() { return b; }
-
-Rectangle::Rectangle(double width, double height)
-{
-	if (width > 0 && height > 0) 
-	{
-		a = width;
-		b = height;
-	}
-	else
-	{
-		throw invalid_argument{ "Нельзя" };
-	}
-}
-double Rectangle::get_area() { return get_composition(); }
-double Rectangle::get_perimeter()
-{
-	return (get_b() * 2) + (get_a() * 2);
-}
